# apps/api/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pickle Go API
  description: |
    Pickle Go is a web-first pickleball event management platform.
    This API provides endpoints for user authentication, event management, and registration.
  version: 1.0.0
  contact:
    name: Pickle Go Team
    url: https://picklego.tw
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.picklego.tw/api/v1
    description: Production server
...
paths:
  /auth/line/callback:
    post:
      tags:
        - Auth
      summary: Line Login Callback
      description: Exchange Line authorization code for JWT tokens
      operationId: lineCallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineCallbackRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh Token
      description: Refresh access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get Current User
      description: Get the authenticated user's profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    LineCallbackRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          description: Authorization code from Line
        state:
          type: string
          description: State parameter for CSRF protection
        redirect_uri:
          type: string
          description: Redirect URI used in the authorization request
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Refresh token
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            access_token:
              type: string
            refresh_token:
              type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        email:
          type: string
          format: email
...